requirements.txt
autopep8==1.6.0
Django==2.2.10
djangorestframework==3.12.4
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
